+++ plugin_info/configuration.php
@@ -21,11 +21,16 @@ if (!isConnect()) {
     include_file('desktop', '404', 'php');
     die();
 }
+$plugin_version = config::byKey('plugin_version', 'Diagral_eOne');
+if (empty($plugin_version)) {
+    $plugin_version = "Development Version";
+}
 ?>
+<p class="font-weight-bold text-right"><b>Version : <?php echo $plugin_version ?></b></p>
 <form class="form-horizontal">
     <fieldset>
         <legend>
-			<i class="fa fa-list-alt"></i> {{Compte Diagral}}
+			<i class="fas fa-key"></i> {{Compte Diagral}}
 		</legend>
         <div class="form-group">
             <label class="col-lg-4 control-label">{{Identifiant}}</label>
@@ -42,7 +47,7 @@ if (!isConnect()) {
     </fieldset>
     <fieldset>
         <legend>
-			<i class="fa fa-list-alt"></i> {{Configuration}}
+			<i class="fas fa-wrench"></i> {{Configuration}}
 		</legend>
         <div class="form-group">
             <label class="col-lg-4 control-label">{{Nombre de tentatives}}</label>
@@ -65,19 +70,10 @@ if (!isConnect()) {
     </fieldset>
     <fieldset>
         <legend>
-			<i class="fa fa-list-alt"></i> {{Communication par le developpeur}}
+			<i class="fas fa-chart-line"></i> {{Suivi d'installation}}
 		</legend>
-        <div class="alert alert-info">
-            Ce plugin étant gratuit, Jeedom ne propose pas au developpeur le suivi des installations de son plugin.<br/>
-            Ce plugin inclut un système de suivi des installations ayant pour objectif :
-            <ul>
-                <li>Suivre le nombre d'installation du plugin (à titre purement informatif pour le developpeur)</li>
-                <li>Avoir un moyen de communication avec les utilisateurs du plugin (envoi de mail ou message privée sur Community)</li>
-            </ul>
-            <br/>
-            Concernant les communications en provenance du développeur, elle seront <b>faibles</b> et avec <b>objectif de pouvoir échanger sur les nouveautés et évolutions à venir</b>.
-            <br/>
-            Afin d'assurer une totale transparence sur les données collectées, elle s'affiche ci-dessous :
+        <div class="alert">
+            <i>Voir la documentation pour plus d'informations sur ce point.</i>
         </div>
         <div class="form-group">
             <label class="col-lg-4 control-label">{{Activation de l'envoi des informations}}</label>
@@ -98,9 +94,21 @@ if (!isConnect()) {
             </div>
         </div>
         <div class="form-group">
-            <label class="col-lg-4 control-label">{{Votre login Jeedom::Market}}</label>
+            <label class="col-lg-4 control-label">{{Votre login Market}}</label>
             <div class="col-lg-4">
                 <span class="label label-info"><?php echo config::byKey('market::username') ?></span>
+            </div>
+        </div>
+        <div class="form-group">
+            <label class="col-lg-4 control-label">{{Version Jeedom}}</label>
+            <div class="col-lg-4">
+                <span class="label label-info"><?php echo jeedom::version() ?></span>
+            </div>
+        </div>
+        <div class="form-group">
+            <label class="col-lg-4 control-label">{{Version du plugin}}</label>
+            <div class="col-lg-4">
+                <span class="label label-info"><?php echo config::byKey('plugin_version', 'Diagral_eOne') ?></span>
             </div>
         </div>
         <div class="form-group">
@@ -109,10 +117,16 @@ if (!isConnect()) {
                 <input class="configKey form-control" data-l1key="InstallBaseEmailAddr" type="email" pattern="[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,}$" placeholder="Saisissez votre email si vous souhaitez recevoir les communications du developpeur"/>
             </div>
         </div>
+        <div class="form-group">
+            <div class="col-lg-4"></div>
+            <div class="col-lg-4">
+                <button type="button" id="delete_remote_datainfo" class="btn btn-danger btn-lg">Supprimer mes données</button>
+            </div>
+        </div>
     </fieldset>
     <fieldset>
         <legend>
-			<i class="fa fa-list-alt"></i> {{Debug}}
+			<i class="fas fa-spider"></i> {{Debug}}
 		</legend>
         <div class="form-group">
             <label class="col-lg-4 control-label">{{Verbose}}</label>
@@ -156,6 +170,30 @@ if (!isConnect()) {
                 }
             });
         }
+
+        // Supprime les données utilisateurs présente sur le cloud Developpeur
+        $('#delete_remote_datainfo').click( function() {
+            $.ajax({// fonction permettant de faire de l'ajax
+                type: "POST", // methode de transmission des données au fichier php
+                url: "plugins/Diagral_eOne/core/ajax/Diagral_eOne.ajax.php", // url du fichier php
+                data: {
+                    action: "delete_remote_datainfo",
+                },
+                dataType: 'json',
+                error: function (request, status, error) {
+                    handleAjaxError(request, status, error);
+                },
+                success: function (data) { // si l'appel a bien fonctionné
+                    if (data.state != 'ok') {
+                        $('#div_alert').showAlert({message: data.result, level: 'danger'});
+                        return;
+                    }
+                    //$('#ul_plugin .li_plugin[data-plugin_id=Diagral_eOne]').click();
+                    // Recharge la page après execution
+                    location.reload(true);
+                }
+            });
+        });
     </script>
 
   </fieldset>